move `get_static_layers_data_access()` into nisar.products.utils
